+++ BE/girin-grim/src/main/java/com/starshop/giringrim/payment/controller/PaymentController.java
@@ -1,6 +1,8 @@
-package com.starshop.giringrim.payment;
+package com.starshop.giringrim.payment.controller;
 
-import com.starshop.giringrim.funding.dto.FundingRespDtos;
+import com.starshop.giringrim.payment.dto.PaymentReqDtos;
+import com.starshop.giringrim.payment.dto.PaymentRespDtos;
+import com.starshop.giringrim.payment.service.PaymentService;
 import com.starshop.giringrim.utils.common.ApiResponseGenerator;
 import com.starshop.giringrim.utils.security.UserDetailsImpl;
 import jakarta.validation.Valid;
@@ -9,8 +11,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
