+++ BE/girin-grim/src/main/java/com/starshop/giringrim/payment/repository/PaymentRepository.java
@@ -1,5 +1,6 @@
-package com.starshop.giringrim.payment;
+package com.starshop.giringrim.payment.repository;
 
+import com.starshop.giringrim.payment.entity.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
